requirements.txt
sqlalchemy
pydantic_settings
alembic
asyncpg
psycopg2
fastapi
PyJWT
websockets